requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
undetected-chromedriver==3.5.4
cloudscraper==1.2.71
fake-useragent==1.4.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
schedule==1.2.0
webdriver-manager==4.0.1
psycopg2-binary==2.9.9
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pytz==2023.3.post1

# New dependencies for ultra-advanced features
dnspython==2.4.2
asyncio==3.4.3
scikit-learn==1.3.2  # For ML-based proxy selection
twocaptcha==0.0.3  # Already included in existing setup
pathlib==1.0.1

# MCP Server dependencies
aiohttp==3.9.1